requirements.txt
# Extracción y Carga
requests==2.31.0
pandas==2.2.0
pyarrow==15.0.0
google-cloud-bigquery==3.17.0
python-dotenv==1.2.2

# Transformación
dbt-bigquery==1.7.5

# Orquestación y Conexiones (Idealmente para tu imagen de Docker)
apache-airflow==2.8.1
apache-airflow-providers-google==10.13.0